requirements.txt
Django==4.2.10
django-environ==0.11.2
psycopg2-binary==2.9.9
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
celery==5.3.6
redis==5.0.1
flower==2.0.1
gunicorn==21.2.0
whitenoise==6.6.0
django-filter==23.5
django-redis==5.4.0
python-json-logger==2.0.7
requests==2.31.0
